requirements.txt
fastapi
uvicorn
jinja2
parselmouth
openai-whisper
